% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dyadic_data.r
\name{get_dyadic_data}
\alias{get_dyadic_data}
\title{get_dyadic_data}
\usage{
get_dyadic_data(data="dyadic_dir")
}
\value{
returns a data.frame in the global enviroment.
}
\description{
\code{get_dyadic_data} Returns a panel data formed by Country-dyads from 1816 to 2016.
Beyond the dyads, the dataset also combines variables from Polity, CINC, and MIDS.
The dataset uses COW code as the country ID. Id values for the dyads are also provided.
Finally, one can ask for direct or indirect dyads.
}
\examples{
get_dyadic_data("dyadic_dir")
}
