% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_each_data.r
\name{get_each_data}
\alias{get_each_data}
\title{get_each_data}
\usage{
get_each_data(data="name of the data")
}
\arguments{
\item{\code{data}}{The name of the data one want to download (character). For this function, this paramater
assumes the following values.
\itemize{
\item \code{PolityIV}  Returns the Polity IV dataset
\item \code{COW}  Returns the Country-Code system from the Correlates of War Project
\item \code{CINC} Returns the National Material Capabilities from the Correlates
of War Project
\item \code{UCDP} Returns the UCDP data on Conflict Onset
\item \code{Trade} Returns the dataset onf trade flow from Professor Kristen Gledistch
\item \code{MID} Returns the Dyadic Militarized Interstate Disputes Dataset from Professor Zeev Maoz}}
}
\value{
returns a data.frame automatically saved in the .global enviroment.
}
\description{
\code{get_each_data} downloads a single IR data in your globalenvironment.
}
\details{
PolityIV: Returns the Polity Data.
}
\examples{
get_each_data("PolityIV")
}
