% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_monadic_data.r
\name{get_merged_data}
\alias{get_merged_data}
\title{get_monadic_data}
\usage{
get_monadic_data(data="monadic")
}
\arguments{
\item{data}{a character with one argument: monadic.}
}
\value{
returns a data.frame automatically saved in the global enviroment.
}
\description{
This function saves in your globalenvironment painel monadyc data using country-year as
units of analysis. This merged dataset is generating from a joining some sommonly used
data in IR, such as PolityIV, Cinc, COW, UCDP, TRADE.
}
\examples{
get_monadic_data("monadic")
}
