% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_monadic_data.r
\name{get_monadic_data}
\alias{get_monadic_data}
\title{get_monadic_data}
\usage{
get_monadic_data(data="monadic")
}
\arguments{
\item{\code{data}}{a character with one argument: \code{Monadic}.}
}
\value{
returns a dataset saved in Rstudio global enviroment.
}
\description{
\code{get_each_data} saves in your globalenvironment a new data using country-year as
units of analysis. The dataset is generated from joining other relevant IR datasets, and combaining
them in a monafyc layout. The datasets combined are s PolityIV, Cinc, COW, UCDP onset and Trade flow. See the options for
\code{get_each_data} to understand more about the raw datasets used.
}
\examples{
get_monadic_data("monadic")
}
