% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dyadic_data.r
\name{get_dyadic_data}
\alias{get_dyadic_data}
\title{get_dyadic_data}
\usage{
get_dyadic_data(data="dyadic_dir")
}
\value{
returns a data.frame automatically saved in the global enviroment.
}
\description{
\code{get_dyadic_data} Returns a dyadic skeleton formed by painel
country-year dataset from 1816 to 2016. It also merges this skeleton with
the data from MID, Polity and CINC, using the COW code as the country ID.
}
\examples{
get_monadic_data("dyadic_dir")
}
