% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dyadic_data.r
\name{get_dyadic_data}
\alias{get_dyadic_data}
\title{get_dyadic_data}
\usage{
get_dyadic_data(data="dyadic_dir")
}
\arguments{
\item{data}{a character with two argument: dyadic_dir and dyadic_und.}

\item{dyadic_dir}{returns a direct dyadic data}

\item{dyadic_und}{returns a undirect dyadic data}
}
\value{
returns a data.frame automatically saved in the global enviroment.
}
\description{
This function saves in your globalenvironment painel dyadic data using country-year as
units of analysis. This merged dataset is generating from a joining some commonly used
data in IR, such as PolityIV, Cinc, COW, UCDP, TRADE.
}
\examples{
get_monadic_data("dyadic_dir")
}
