% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_each_data.r
\name{get_each_data}
\alias{get_each_data}
\title{get_each_data}
\usage{
get_each_data(data="name of the data")
}
\arguments{
\item{data}{The name of the data one want to download (character). For this function, this paramater
assumes the following values: PolityIV, COW, CINC, MID, UCDP, Trade, outros.}
}
\value{
returns a data.frame automatically saved in the .global enviroment.
}
\description{
This function saves datasets on IR in your globalenvironment in a easy way.
}
\examples{
get_each_data("PolityIV")
}
