% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_monadic_data.r
\name{get_monadic_data}
\alias{get_monadic_data}
\title{get_monadic_data}
\usage{
get_monadic_data(data="monadic")
}
\arguments{
\item{\code{data}}{a character with one argument: \code{Monadic}.}
}
\value{
returns a data.frame automatically saved in the global enviroment.
}
\description{
\code{get_each_data} saves in your globalenvironment painel monadyc data using country-year as
units of analysis. This merged dataset is generating from a joining some fairly  used
data in IR, such as PolityIV, Cinc, COW, UCDP onset and Trade flow. See the options for
\code{get_each_data} to understand more about these datasets.
}
\examples{
get_monadic_data("monadic")
}
